+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 ---------
 
+v0.4.13
+=======
+- GraphQL: use async methods for query resolvers. Fixed using mutation methods and passing variables.
+
+v0.4.13
+=======
+- GraphQL basic support (See examples/graphql)
+
 v0.4.12
 =======
 - Fixed fragmentation for fire and forget